fix/data_manager: fix Put and Post handling in data manager
This prevents a group from going out of sync in the rare case where
half of the nodes receive one Put/Post request first, and the other
half receive a different Put/Post request for the same data chunk first.
Previously, every node would handle and put in the chunk store the data
it got first. With the new approach, this situation would simply lead to
both requests being refused, leaving the data intact and in its previous
state:

When the nodes receive a request, they don't immediately write to the
chunk store, and instead cache the request in memory. Each node sends,
as a message from the group to itself, the _first_ version of the
data that it received. Only if that message accumulates, i.e. if a
quorum of nodes agrees on which request to choose, they write it to
disk.