chore/ci: explicit release builds
Refactors the process to have explicit release builds for macOS and
Windows. Linux was already doing this anyway because it was using musl,
which uses a slightly different build process.

The release builds completely clear out any cache so that everything is
built from fresh source. However, this only applies when we're building
the master branch. For a PR, it won't clear out the cache.