refactor(ci): Refactor of GHA CI files
Updated version of the checkout action being used to v2.
Tidied and condensed the caching.
Added a checkout depth of '0' when adding a tag to the repo to
ensure all commits are retrieved (though the publish part of the
code is actually commented out at the moment).
Switched the code coverage to it's own job so it can run in
parallel, and set it to only run against PRs.