Added flag to clear data of previous running node
Also removed hard coded integral constants like `ARGS[18]` (hardly
maintainable, you have to count elements in ARGS array to check what
an index stands for)

The situation is better now (no index values to manage anymore) but not
perfect. The problem is that I don't understand why arguments are
analyzed this way. This forces the creation of an array that duplicates
the structure used to define the arguments.
Normally with structopt we don't need to do that.

Also corrected some clippy and fmt errors.